Ticketmaster fix
This fix includes one month period of date, various image return names
and some issues on sending response from server

+++ controllers/attractions.js
@@ -388,7 +388,7 @@ exports.getAttractionsInRadiusWithLongLat = async (req, res, next) => {
 
    const longitute = parseFloat(lng)
    const latitude = parseFloat(lat)
-   const rad = parseInt(radius)
+   const rad = parseFloat(radius)
 
    console.log(`${longitute} ${latitude} ${rad}`)
